Update DataStore API to work with new spec
The main methods have been renamed to "store" and "retrieve". Retrieve operations
take a reference and return a stream as before. Store operations take a stream and
return a reference. A Worker is not responsible for generating a reference. The relative
location of the object will be set by the DataStore implementation configuration.
For this reason, the "resolve" method has been removed.