SCMOD-11585: Attempt to fix random test failure
https://portal.digitalsafe.net/browse/SCMOD-11585

This test is failing at random with the below error. A previous attempt to fix it increased the timeout from 1 second to 5 seconds. This is an attempt to fix it by setting up the mocks before starting the publisher thread.

org.mockito.exceptions.misusing.UnfinishedStubbingException: 

Unfinished stubbing detected here:
-> at com.hpe.caf.worker.queue.rabbit.RabbitWorkerQueuePublisherTest.testHandlePublish(RabbitWorkerQueuePublisherTest.java:78)

E.g. thenReturn() may be missing.
Examples of correct stubbing:
    when(mock.isOk()).thenReturn(true);
    when(mock.isOk())).thenThrow(exception);
    doThrow(exception).when(mock).someVoidMethod();
Hints:
 1. missing thenReturn()
 2. you are trying to stub a final method, which is not supported
 3. you are stubbing the behaviour of another mock inside before 'thenReturn' instruction is completed

	at com.hpe.caf.worker.queue.rabbit.RabbitWorkerQueuePublisherTest.testHandlePublish(RabbitWorkerQueuePublisherTest.java:78